Fix file import

+++ Headache Companion/Views/SettingsView.swift
@@ -312,45 +312,48 @@ struct SettingsView: View {
                     .fileImporter(isPresented: $showingMedFilePicker, allowedContentTypes: [.json]) { result in
                         do {
                             let fileURL = try result.get()
-                            if let allMedImportData = try? JSONSerialization.jsonObject(with: Data(contentsOf: fileURL), options: .allowFragments) as? [[String: Any]]
-                            {
-                                // Delete all current data if user requested
-                                if overwriteMedHistory {
-                                    deleteAllMedHistory()
-                                }
-                                
-                                for hcMedHistory in allMedImportData {
-                                    var newStartDate: Date? = nil
-                                    var newStopDate: Date? = nil
-                                    if hcMedHistory["startDate"] as? TimeInterval != nil {
-                                        newStartDate = Date(timeIntervalSince1970: hcMedHistory["startDate"] as! TimeInterval)
-                                    }
-                                    if hcMedHistory["stopDate"] as? TimeInterval != nil {
-                                        newStopDate = Date(timeIntervalSince1970: hcMedHistory["stopDate"] as! TimeInterval)
+                            if fileURL.startAccessingSecurityScopedResource() {
+                                if let allMedImportData = try? JSONSerialization.jsonObject(with: Data(contentsOf: fileURL), options: .allowFragments) as? [[String: Any]]
+                                {
+                                    // Delete all current data if user requested
+                                    if overwriteMedHistory {
+                                        deleteAllMedHistory()
                                     }
                                     
-                                    // Check if the MedHistory being imported already exists
-                                    if !overwriteMedHistory {
-                                        if let duplicateData = medHistory.first(where: { $0.id == hcMedHistory["id"] as? String }) {
-                                            viewContext.delete(duplicateData)
+                                    for hcMedHistory in allMedImportData {
+                                        var newStartDate: Date? = nil
+                                        var newStopDate: Date? = nil
+                                        if hcMedHistory["startDate"] as? TimeInterval != nil {
+                                            newStartDate = Date(timeIntervalSince1970: hcMedHistory["startDate"] as! TimeInterval)
                                         }
+                                        if hcMedHistory["stopDate"] as? TimeInterval != nil {
+                                            newStopDate = Date(timeIntervalSince1970: hcMedHistory["stopDate"] as! TimeInterval)
+                                        }
+                                        
+                                        // Check if the MedHistory being imported already exists
+                                        if !overwriteMedHistory {
+                                            if let duplicateData = medHistory.first(where: { $0.id == hcMedHistory["id"] as? String }) {
+                                                viewContext.delete(duplicateData)
+                                            }
+                                        }
+                                        
+                                        let newMedHistory = MedHistory(context: viewContext)
+                                        newMedHistory.startDate = newStartDate
+                                        newMedHistory.startDate = newStopDate
+                                        newMedHistory.id = hcMedHistory["id"] as? String
+                                        newMedHistory.name = hcMedHistory["name"] as! String
+                                        newMedHistory.amount = hcMedHistory["amount"] as! Int32
+                                        newMedHistory.dose = hcMedHistory["dose"] as! String
+                                        newMedHistory.frequency = hcMedHistory["frequency"] as! String
+                                        newMedHistory.effective = Effectiveness(rawValue: hcMedHistory["effective"] as? Int16 ?? 2)!
+                                        newMedHistory.notes = hcMedHistory["notes"] as! String
+                                        newMedHistory.type = hcMedHistory["type"] as! String
+                                        newMedHistory.sideEffects = hcMedHistory["sideEffects"] as? Set<String>
                                     }
-                                    
-                                    let newMedHistory = MedHistory(context: viewContext)
-                                    newMedHistory.startDate = newStartDate
-                                    newMedHistory.startDate = newStopDate
-                                    newMedHistory.id = hcMedHistory["id"] as? String
-                                    newMedHistory.name = hcMedHistory["name"] as! String
-                                    newMedHistory.amount = hcMedHistory["amount"] as! Int32
-                                    newMedHistory.dose = hcMedHistory["dose"] as! String
-                                    newMedHistory.frequency = hcMedHistory["frequency"] as! String
-                                    newMedHistory.effective = Effectiveness(rawValue: hcMedHistory["effective"] as? Int16 ?? 2)!
-                                    newMedHistory.notes = hcMedHistory["notes"] as! String
-                                    newMedHistory.type = hcMedHistory["type"] as! String
-                                    newMedHistory.sideEffects = hcMedHistory["sideEffects"] as? Set<String>
+                                    saveData()
                                 }
-                                saveData()
                             }
+                            fileURL.stopAccessingSecurityScopedResource()
                         } catch {
                             print("Failed to import data: \(error.localizedDescription)")
                         }
